backend/payments/requirements.txt
psycopg2-binary==2.9.9
PyJWT==2.8.0
